chore: refactor Containers into TestTenantDb
