+++ test/realtime/messages_test.exs
@@ -8,7 +8,7 @@ defmodule Realtime.MessagesTest do
   alias Realtime.Tenants.Repo
 
   setup do
-    tenant = Containers.checkout_tenant(run_migrations: true)
+    tenant = TestTenantDb.checkout_tenant(run_migrations: true)
     {:ok, conn} = Database.connect(tenant, "realtime_test", :stop)
 
     date_start = Date.utc_today() |> Date.add(-10)
