+++ test/realtime/tenants/janitor/maintenance_task_test.exs
@@ -7,7 +7,7 @@ defmodule Realtime.Tenants.Janitor.MaintenanceTaskTest do
   alias Realtime.Tenants.Repo
 
   setup do
-    tenant = Containers.checkout_tenant(run_migrations: true)
+    tenant = TestTenantDb.checkout_tenant(run_migrations: true)
     # Warm cache to avoid Cachex and Ecto.Sandbox ownership issues
     Realtime.Tenants.Cache.update_cache(tenant)
