+++ test/realtime/tenants/schema_test.exs
@@ -9,7 +9,7 @@ defmodule Realtime.Tenants.SchemaTest do
   alias Realtime.Database
 
   setup do
-    tenant = Containers.checkout_tenant(run_migrations: true)
+    tenant = TestTenantDb.checkout_tenant(run_migrations: true)
     {:ok, settings} = Database.from_tenant(tenant, "realtime_test", :stop)
     opts = settings |> Map.from_struct() |> Keyword.new()
