+++ test/realtime_web/channels/realtime_channel/broadcast_handler_test.exs
@@ -458,7 +458,7 @@ defmodule RealtimeWeb.RealtimeChannel.BroadcastHandlerTest do
   end
 
   defp initiate_tenant(context) do
-    tenant = Containers.checkout_tenant(run_migrations: true)
+    tenant = TestTenantDb.checkout_tenant(run_migrations: true)
 
     # Warm cache to avoid Cachex and Ecto.Sandbox ownership issues
     Realtime.Tenants.Cache.update_cache(tenant)
