+++ test/realtime_web/channels/realtime_channel/presence_handler_test.exs
@@ -774,7 +774,7 @@ defmodule RealtimeWeb.RealtimeChannel.PresenceHandlerTest do
   end
 
   defp initiate_tenant(context) do
-    tenant = Containers.checkout_tenant(run_migrations: true)
+    tenant = TestTenantDb.checkout_tenant(run_migrations: true)
     # Warm cache to avoid Cachex and Ecto.Sandbox ownership issues
     Realtime.Tenants.Cache.update_cache(tenant)
