+++ test/realtime_web/channels/realtime_channel_replication_ready_test.exs
@@ -9,7 +9,7 @@ defmodule RealtimeWeb.RealtimeChannelReplicationReadyTest do
   setup :set_mimic_from_context
 
   setup do
-    tenant = Containers.checkout_tenant(run_migrations: true)
+    tenant = TestTenantDb.checkout_tenant(run_migrations: true)
     Realtime.Tenants.Cache.update_cache(tenant)
     {:ok, tenant: tenant}
   end
