+++ test/realtime_web/channels/realtime_channel_test.exs
@@ -15,7 +15,7 @@ defmodule RealtimeWeb.RealtimeChannelTest do
   alias RealtimeWeb.UserSocket
 
   setup do
-    tenant = Containers.checkout_tenant(run_migrations: true)
+    tenant = TestTenantDb.checkout_tenant(run_migrations: true)
     {:ok, db_conn} = Realtime.Database.connect(tenant, "realtime_test", :stop)
     Integrations.setup_postgres_changes(db_conn)
     GenServer.stop(db_conn)
