+++ test/realtime_web/dashboard/tenant_migrations_test.exs
@@ -16,7 +16,7 @@ defmodule RealtimeWeb.Dashboard.TenantMigrationsTest do
       Application.delete_env(:realtime, :dashboard_credentials)
     end)
 
-    tenant = Containers.checkout_tenant(run_migrations: true)
+    tenant = TestTenantDb.checkout_tenant(run_migrations: true)
     conn = using_basic_auth(build_conn(), "user", "pass")
 
     %{tenant: tenant, conn: conn}
