--- test/support/containers.ex
@@ -1,521 +1,0 @@
-defmodule Containers do
-  alias Extensions.PostgresCdcRls
-  alias Realtime.Tenants.Connect
-  alias Containers.Container
-  alias Containers.ExternalTarget
-  alias Realtime.Database
-  alias Realtime.Env
-  alias Realtime.Tenants
-  alias Realtime.Tenants.Migrations
-
-  use GenServer
-
-  # External-tenant-DB mode (USE_EXTERNAL_TENANT_DB=true): instead of starting
-  # per-test supabase/postgres containers, point DB-backed tenant tests at one
-  # or more already-running, external Postgres-wire-compatible servers (e.g.
-  # Multigres) on EXTERNAL_TENANT_DB_PORTS, host fixed at 127.0.0.1. Each
-  # configured port is an independent DB — the pool hands out exactly one per
-  # concurrent test, same checkout/checkin contract as the container pool.
-  # Unset, behavior is unchanged. Uses Env.get_boolean/2 for consistency with
-  # the rest of the codebase (e.g. DB_SSL) — accepts true/1/false/0
-  # (case-insensitive), raises on anything else rather than silently treating
-  # a typo as false.
-  def external_tenant_db?, do: Env.get_boolean("USE_EXTERNAL_TENANT_DB", false)
-
-  # EXTERNAL_TENANT_DB_PORTS is a comma-separated list, one port per
-  # independent external DB (e.g. one per Multigres cluster). EXTERNAL_TENANT_DB_PORT
-  # (singular) is kept as an alias for a single-port list.
-  def external_tenant_db_ports! do
-    case presence(System.get_env("EXTERNAL_TENANT_DB_PORTS")) || presence(System.get_env("EXTERNAL_TENANT_DB_PORT")) do
-      nil ->
-        raise "USE_EXTERNAL_TENANT_DB=true requires EXTERNAL_TENANT_DB_PORTS (comma-separated) or EXTERNAL_TENANT_DB_PORT to be set"
-
-      value ->
-        value
-        |> String.split(",")
-        |> Enum.map(&String.trim/1)
-        |> Enum.map(&String.to_integer/1)
-    end
-  end
-
-  # Treats an explicitly-empty env var the same as unset, so e.g.
-  # EXTERNAL_TENANT_DB_PORTS="" falls through to EXTERNAL_TENANT_DB_PORT (or
-  # the clear "not set" error) instead of raising a cryptic ArgumentError out
-  # of String.to_integer("").
-  defp presence(nil), do: nil
-  defp presence(""), do: nil
-  defp presence(value), do: value
-
-  defp image, do: System.get_env("POSTGRES_IMAGE", "supabase/postgres:17.6.1.127")
-
-  # Pull image if not available
-  def pull do
-    case System.cmd("docker", ["image", "inspect", image()]) do
-      {_, 0} ->
-        :ok
-
-      _ ->
-        IO.puts("Pulling image #{image()}. This might take a while...")
-        {_, 0} = System.cmd("docker", ["pull", image()])
-    end
-  end
-
-  def start_container(), do: GenServer.call(__MODULE__, :start_container, 10_000)
-  def start_external_target(), do: GenServer.call(__MODULE__, :start_external_target, 10_000)
-  def port(), do: GenServer.call(__MODULE__, :port, 10_000)
-
-  def start_link(max_cases), do: GenServer.start_link(__MODULE__, max_cases, name: __MODULE__)
-
-  def init(max_cases) do
-    # In external mode no containers are ever started, so skip the docker
-    # probe — `docker ps` would otherwise crash Containers.init/1
-    # (ErlangError if docker isn't installed, MatchError if the daemon is
-    # down) before any test can run, defeating the point of this mode. The
-    # port-range computation below stays docker-free either way (it only
-    # excludes already-used ports from existing_containers) — it's still
-    # needed in external mode as the fallback Containers.port/0 uses for
-    # Generators.tenant_fixture/1's default db_port when a test doesn't pass
-    # one explicitly.
-    existing_containers = if external_tenant_db?(), do: [], else: existing_containers("realtime-test-*")
-    ports = for {_, port} <- existing_containers, do: port
-
-    partition = System.get_env("MIX_TEST_PARTITION", "1") |> String.to_integer()
-    total_partitions = System.get_env("MIX_TEST_TOTAL_PARTITIONS", "4") |> String.to_integer()
-    all_ports = 6500..9000
-    range_size = div(Enum.count(all_ports), total_partitions)
-
-    available_ports =
-      all_ports |> Enum.slice((partition - 1) * range_size, range_size) |> Enum.shuffle() |> Kernel.--(ports)
-
-    external_ports = if external_tenant_db?(), do: external_tenant_db_ports!(), else: []
-
-    state = %{
-      existing_containers: existing_containers,
-      ports: available_ports,
-      external_ports: external_ports,
-      external_port_owners: %{}
-    }
-
-    {:ok, state, {:continue, {:pool, max_cases}}}
-  end
-
-  def handle_continue({:pool, max_cases}, state) do
-    # One ExternalTarget worker per configured external port (each just
-    # claims an already-running DB, no docker to start/pull) in external
-    # mode; otherwise one Container worker per max concurrent test plus a
-    # small buffer.
-    {worker_module, size} =
-      if external_tenant_db?() do
-        {Containers.ExternalTarget, length(state.external_ports)}
-      else
-        {Containers.Container, max_cases + 2}
-      end
-
-    {:ok, _pid} =
-      :poolboy.start_link(
-        [
-          strategy: :fifo,
-          name: {:local, Containers.Pool},
-          size: size,
-          max_overflow: 0,
-          worker_module: worker_module
-        ],
-        []
-      )
-
-    {:noreply, state}
-  end
-
-  def handle_call(:port, _from, state) do
-    [port | ports] = state.ports
-    {:reply, port, %{state | ports: ports}}
-  end
-
-  def handle_call(:start_container, _from, state) do
-    case state.existing_containers do
-      [{name, port} | rest] ->
-        {:reply, {:ok, name, port}, %{state | existing_containers: rest}}
-
-      [] ->
-        {name, port, ports} = start_available_container(state.ports)
-        {:reply, {:ok, name, port}, %{state | ports: ports}}
-    end
-  end
-
-  # Monitors the requesting ExternalTarget worker and tracks which port it
-  # holds, so a crashed worker's port is reclaimed (see handle_info :DOWN
-  # below) instead of being lost forever. Pool size equals
-  # length(external_ports) exactly (no spare), so a poolboy-restarted worker
-  # can otherwise permanently exhaust the list after just one crash.
-  def handle_call(:start_external_target, {pid, _tag}, state) do
-    case state.external_ports do
-      [port | rest] ->
-        Process.monitor(pid)
-        state = %{state | external_ports: rest, external_port_owners: Map.put(state.external_port_owners, pid, port)}
-        {:reply, port, state}
-
-      [] ->
-        # Momentary exhaustion — e.g. a replacement worker asked before the
-        # dead worker's :DOWN reclaimed its port. The caller retries.
-        {:reply, {:error, :no_external_ports_available}, state}
-    end
-  end
-
-  def handle_info({:DOWN, _ref, :process, pid, _reason}, state) do
-    case Map.pop(state.external_port_owners, pid) do
-      {nil, _owners} ->
-        {:noreply, state}
-
-      {port, owners} ->
-        {:noreply, %{state | external_ports: [port | state.external_ports], external_port_owners: owners}}
-    end
-  end
-
-  defp start_available_container(ports, attempts \\ 5)
-
-  defp start_available_container([], _attempts), do: raise("Containers: no ports left to start a container")
-  defp start_available_container(_ports, 0), do: raise("Containers: exhausted retries starting a container")
-
-  defp start_available_container([port | ports], attempts) do
-    name = "realtime-test-#{random_string(12)}"
-
-    case docker_run(name, port) do
-      {_, 0} -> {name, port, ports}
-      {_output, _code} -> start_available_container(ports, attempts - 1)
-    end
-  end
-
-  defp random_string(length) do
-    :crypto.strong_rand_bytes(length)
-    |> Base.url_encode64()
-    |> binary_part(0, length)
-  end
-
-  def initialize(external_id) do
-    name = "realtime-tenant-test-#{external_id}"
-
-    port =
-      case existing_containers(name) do
-        [{^name, port}] ->
-          port
-
-        [] ->
-          port = 5500
-          docker_run!(name, port)
-          port
-      end
-
-    check_container_ready(name)
-
-    opts = %{external_id: external_id, name: external_id, port: port, jwt_secret: "secure_jwt_secret"}
-    tenant = Generators.tenant_fixture(opts)
-
-    Migrations.run_migrations(tenant)
-    {:ok, pid} = Database.connect(tenant, "realtime_test", :stop)
-    :ok = Tenants.create_messages_partitions(pid)
-    Process.exit(pid, :normal)
-
-    tenant
-  end
-
-  @doc "Return port for a container (or an external tenant DB) that can be used"
-  def checkout() do
-    with container when is_pid(container) <- :poolboy.checkout(Containers.Pool, true, 5_000),
-         port <- port_module().port(container) do
-      # Automatically checkin the container at the end of the test
-      ExUnit.Callbacks.on_exit(fn -> :poolboy.checkin(Containers.Pool, container) end)
-
-      {:ok, port}
-    else
-      _ -> {:error, "failed to checkout a container"}
-    end
-  end
-
-  # In external-tenant-db mode the tenant DB is the external server's
-  # pre-existing `postgres` database; there's nothing to create, and some
-  # Postgres-wire-compatible servers don't implement CREATE DATABASE (or
-  # mishandle Ecto's zero-column existence probe), so skip this entirely.
-  defp storage_up!(tenant) do
-    if external_tenant_db?() do
-      :ok
-    else
-      do_storage_up!(tenant)
-    end
-  end
-
-  defp do_storage_up!(tenant) do
-    {:ok, db_settings} = Database.from_tenant(tenant, "realtime_test", :stop)
-
-    settings =
-      db_settings
-      |> Map.from_struct()
-      |> Keyword.new()
-
-    case Ecto.Adapters.Postgres.storage_up(settings) do
-      :ok -> :ok
-      {:error, :already_up} -> :ok
-      _ -> raise "Failed to create database"
-    end
-  end
-
-  def checkout_tenant(opts \\ []), do: do_checkout_tenant(opts, :sandbox)
-  def checkout_tenant_unboxed(opts \\ []), do: do_checkout_tenant(opts, :unboxed)
-
-  # Acquire a tenant database for one test — a pooled supabase/postgres
-  # container, or (in external mode) one of the pre-configured external DBs.
-  # Either way it's a real pool checkout, released via the returned checkin
-  # function once the caller is done.
-  defp acquire_tenant_db do
-    case :poolboy.checkout(Containers.Pool, true, 5_000) do
-      container when is_pid(container) ->
-        {:ok, port_module().port(container), fn -> :poolboy.checkin(Containers.Pool, container) end}
-
-      _ ->
-        :error
-    end
-  end
-
-  defp port_module, do: if(external_tenant_db?(), do: ExternalTarget, else: Container)
-
-  defp do_checkout_tenant(opts, mode) do
-    with {:ok, port, checkin} <- acquire_tenant_db() do
-      tenant = repo_run(mode, fn -> Generators.tenant_fixture(%{port: port, migrations_ran: 0}) end)
-
-      run_migrations? = Keyword.get(opts, :run_migrations, false)
-
-      {:ok, settings} = Database.from_tenant(tenant, "realtime_test", :stop)
-      settings = %{settings | max_restarts: 0, ssl: false}
-      {:ok, conn} = Database.connect_db(settings)
-
-      try do
-        reset_realtime_schema!(settings)
-        storage_up!(tenant)
-
-        RateCounterHelper.stop(tenant.external_id)
-
-        ExUnit.Callbacks.on_exit(fn ->
-          if connect_pid = Connect.whereis(tenant.external_id) do
-            supervisor = {:via, PartitionSupervisor, {Realtime.Tenants.Connect.DynamicSupervisor, tenant.external_id}}
-
-            DynamicSupervisor.terminate_child(supervisor, connect_pid)
-          end
-
-          try do
-            PostgresCdcRls.handle_stop(tenant.external_id, 5_000)
-          catch
-            _, _ -> :ok
-          end
-
-          if mode == :unboxed do
-            repo_run(:unboxed, fn -> Realtime.Api.delete_tenant_by_external_id(tenant.external_id) end)
-          end
-
-          checkin.()
-        end)
-
-        if run_migrations? do
-          case run_migrations(tenant) do
-            {:ok, count} ->
-              :ok = Tenants.create_messages_partitions(conn)
-
-              {:ok, tenant} =
-                repo_run(mode, fn ->
-                  Realtime.Api.update_tenant_by_external_id(tenant.external_id, %{migrations_ran: count})
-                end)
-
-              if mode == :sandbox, do: Realtime.Tenants.Cache.invalidate_tenant_cache(tenant.external_id)
-
-              tenant
-
-            error ->
-              raise "Failed to run migrations: #{inspect(error)}"
-          end
-        else
-          tenant
-        end
-      after
-        GenServer.stop(conn)
-      end
-    else
-      :error -> {:error, "failed to checkout a container"}
-    end
-  end
-
-  defp repo_run(:unboxed, fun), do: Ecto.Adapters.SQL.Sandbox.unboxed_run(Realtime.Repo, fun)
-  defp repo_run(:sandbox, fun), do: fun.()
-
-  defp reset_realtime_schema!(settings, attempts \\ 5) do
-    {:ok, admin_conn} =
-      Postgrex.start_link(
-        hostname: settings.hostname,
-        port: settings.port,
-        database: settings.database,
-        username: "supabase_admin",
-        password: settings.password
-      )
-
-    # Mirrors supabase/postgres migrations
-    try do
-      Postgrex.query!(admin_conn, "DROP PUBLICATION IF EXISTS supabase_realtime_test", [])
-      Postgrex.query!(admin_conn, "DROP SCHEMA IF EXISTS realtime CASCADE", [])
-      Postgrex.query!(admin_conn, "CREATE SCHEMA realtime", [])
-
-      Postgrex.query!(admin_conn, "GRANT USAGE ON SCHEMA realtime TO postgres", [])
-      Postgrex.query!(admin_conn, "GRANT ALL ON ALL TABLES IN SCHEMA realtime TO postgres, dashboard_user", [])
-      Postgrex.query!(admin_conn, "GRANT ALL ON ALL SEQUENCES IN SCHEMA realtime TO postgres, dashboard_user", [])
-      Postgrex.query!(admin_conn, "GRANT ALL ON ALL ROUTINES IN SCHEMA realtime TO postgres, dashboard_user", [])
-
-      Postgrex.query!(
-        admin_conn,
-        "ALTER DEFAULT PRIVILEGES FOR ROLE supabase_admin IN SCHEMA realtime GRANT ALL ON TABLES TO postgres, dashboard_user",
-        []
-      )
-
-      Postgrex.query!(
-        admin_conn,
-        "ALTER DEFAULT PRIVILEGES FOR ROLE supabase_admin IN SCHEMA realtime GRANT ALL ON SEQUENCES TO postgres, dashboard_user",
-        []
-      )
-
-      Postgrex.query!(
-        admin_conn,
-        "ALTER DEFAULT PRIVILEGES FOR ROLE supabase_admin IN SCHEMA realtime GRANT ALL ON ROUTINES TO postgres, dashboard_user",
-        []
-      )
-
-      Postgrex.query!(admin_conn, "GRANT USAGE ON SCHEMA realtime TO anon, authenticated, service_role", [])
-      Postgrex.query!(admin_conn, "GRANT ALL ON SCHEMA realtime TO supabase_realtime_admin WITH GRANT OPTION", [])
-    rescue
-      # Retry in case of OrioleDB OTablesMetaTranche LWLock
-      e in [Postgrex.Error, DBConnection.ConnectionError] ->
-        GenServer.stop(admin_conn)
-
-        if attempts > 1 do
-          Process.sleep(500)
-          reset_realtime_schema!(settings, attempts - 1)
-        else
-          reraise e, __STACKTRACE__
-        end
-    after
-      if Process.alive?(admin_conn), do: GenServer.stop(admin_conn)
-    end
-  end
-
-  def stop_containers() do
-    {list, 0} = System.cmd("docker", ["ps", "-a", "--format", "{{.Names}}", "--filter", "name=realtime-test-*"])
-
-    for name <- String.split(list, "\n", trim: true) do
-      System.cmd("docker", ["rm", "-f", name])
-    end
-  end
-
-  def stop_container(external_id) do
-    name = "realtime-tenant-test-#{external_id}"
-    System.cmd("docker", ["rm", "-f", name])
-  end
-
-  defp existing_containers(pattern) do
-    {containers, 0} = System.cmd("docker", ["ps", "--format", "{{json .}}", "--filter", "name=#{pattern}"])
-
-    containers
-    |> String.split("\n", trim: true)
-    |> Enum.map(fn container ->
-      container = Jason.decode!(container)
-      # Ports" => "0.0.0.0:6445->5432/tcp, [::]:6445->5432/tcp"
-      regex = ~r/(?<=:)\d+(?=->)/
-
-      [port] =
-        Regex.scan(regex, container["Ports"])
-        |> List.flatten()
-        |> Enum.uniq()
-
-      {container["Names"], String.to_integer(port)}
-    end)
-  end
-
-  defp check_container_ready(name, attempts \\ 50)
-  defp check_container_ready(name, 0), do: raise("Container #{name} is not ready")
-
-  defp check_container_ready(name, attempts) do
-    case System.cmd("docker", ["exec", name, "pg_isready", "-p", "5432", "-h", "localhost"]) do
-      {_, 0} ->
-        :ok
-
-      {_, _} ->
-        Process.sleep(500)
-        check_container_ready(name, attempts - 1)
-    end
-  end
-
-  # This exists so we avoid using an external process on Realtime.Tenants.Migrations
-  defp run_migrations(tenant) do
-    %{extensions: [%{settings: settings} | _]} = tenant
-    {:ok, settings} = Database.from_settings(settings, "realtime_migrations", :stop)
-
-    [
-      hostname: settings.hostname,
-      port: settings.port,
-      database: settings.database,
-      password: settings.password,
-      username: settings.username,
-      pool_size: settings.pool_size,
-      backoff_type: settings.backoff_type,
-      socket_options: settings.socket_options,
-      parameters: [application_name: settings.application_name],
-      ssl: settings.ssl
-    ]
-    |> Realtime.Repo.with_dynamic_repo(fn repo ->
-      try do
-        opts = [all: true, prefix: "realtime", dynamic_repo: repo, log: false]
-        migrations = Realtime.Tenants.Migrations.migrations()
-        Ecto.Migrator.run(Realtime.Repo, migrations, :up, opts)
-
-        {:ok, length(migrations)}
-      rescue
-        error ->
-          {:error, error}
-      end
-    end)
-  end
-
-  defp docker_run!(name, port) do
-    {_, 0} = docker_run(name, port)
-  end
-
-  defp docker_run(name, port) do
-    initdb_sh = Path.expand("../../dev/postgres/za-permit-supabase-admin.sh", __DIR__)
-    initdb_sql = Path.expand("../../dev/postgres/zb-supabase-schema.sql", __DIR__)
-
-    System.cmd(
-      "docker",
-      [
-        "run",
-        "-d",
-        "--rm",
-        "--name",
-        name,
-        "-e",
-        "POSTGRES_HOST=/var/run/postgresql",
-        "-e",
-        "POSTGRES_PASSWORD=postgres",
-        "-v",
-        "#{initdb_sh}:/docker-entrypoint-initdb.d/za-permit-supabase-admin.sh",
-        "-v",
-        "#{initdb_sql}:/docker-entrypoint-initdb.d/zb-supabase-schema.sql",
-        "-p",
-        "#{port}:5432",
-        image(),
-        "postgres",
-        "-c",
-        "config_file=/etc/postgresql/postgresql.conf",
-        "-c",
-        "wal_keep_size=32MB",
-        "-c",
-        "max_wal_size=1GB",
-        "-c",
-        "max_slot_wal_keep_size=32MB"
-      ],
-      stderr_to_stdout: true
-    )
-  end
-end
