+++ test/test_helper.exs
@@ -2,28 +2,12 @@ start_time = :os.system_time(:millisecond)
 
 alias Realtime.Api
 
-# External-tenant-db mode (USE_EXTERNAL_TENANT_DB=true): each configured
-# port is an independent DB reused across the whole run, so max_cases must
-# equal the number of configured ports exactly, not just be <= it —
-# oversubscribing beyond the port count causes far worse, cascading
-# failures than running serially (concurrent tenant setup, e.g. DROP SCHEMA
-# realtime CASCADE, stomping on each other once demand exceeds supply).
-# max_cases is therefore forced here rather than read from MAX_CASES.
-max_cases =
-  if Containers.external_tenant_db?() do
-    forced_max_cases = length(Containers.external_tenant_db_ports!())
-
-    if System.get_env("MAX_CASES") do
-      IO.puts(
-        "[test_helper.exs] USE_EXTERNAL_TENANT_DB=true: ignoring MAX_CASES, forcing " <>
-          "max_cases to #{forced_max_cases} (the number of configured external ports)."
-      )
-    end
-
-    forced_max_cases
-  else
-    String.to_integer(System.get_env("MAX_CASES", "4"))
-  end
+# Where tenant DBs come from — docker containers (default) or external
+# servers (USE_EXTERNAL_TENANT_DB=true). The backend is resolved exactly
+# once per run, here; everything else dispatches through
+# TestTenantDb.Backend.current().
+backend = TestTenantDb.Backend.resolve!()
+max_cases = backend.max_cases()
 
 repo_config = Application.fetch_env!(:realtime, Realtime.Repo)
 
@@ -79,19 +63,9 @@ ExUnit.start(
 
 max_cases = ExUnit.configuration()[:max_cases]
 
-# In external mode no supabase/postgres image is used, so skip the pull and
-# the container teardown. The Containers GenServer still starts (tests call
-# Containers.port()), but its eager poolboy pool is skipped (see
-# handle_continue in containers.ex).
-unless Containers.external_tenant_db?() do
-  Containers.pull()
+backend.prepare!()
 
-  if System.get_env("REUSE_CONTAINERS") != "true" do
-    Containers.stop_containers()
-  end
-end
-
-{:ok, _pid} = Containers.start_link(max_cases)
+{:ok, _pid} = TestTenantDb.start_link(max_cases)
 
 for tenant <- Api.list_tenants(), do: Api.delete_tenant_by_external_id(tenant.external_id)
